+++ python/datamodel/models/releases.py
@@ -13,50 +13,43 @@
 
 from __future__ import print_function, division, absolute_import
 
-from typing import List
+from datetime import date
+from typing import List, Union
 from pydantic import BaseModel, validator
 
 from ..io.loaders import read_yaml, get_yaml_files
-
+from .base import BaseList
 
 class Release(BaseModel):
-    """ Pydantic model presenting an SDSS release """
+    """ Pydantic model presenting an SDSS release
+
+    Parameters
+    ----------
+    name : str
+        The name of the release
+    description : str
+        A description of the release
+    public : bool
+        Whether the release is public or not
+    release_date : `datetime.date`
+        The date of the release
+
+    Raises
+    ------
+    ValueError
+        when the release name does not start with a valid SDSS release code
+    """
     name: str
     description: str
     public: bool = False
+    release_date: Union[str, date] = 'unreleased'
 
     @validator('name')
     def name_check(cls, value): # pylint: disable=no-self-argument
-        if not value.startswith(('WORK', 'DR', 'MPL', 'IPL')):
+        if not value.startswith(('WORK', 'DR', 'MPL', 'IPL', 'EDR')):
             raise ValueError('release name must start with WORK, DR, MPL, or IPL.')
         return value
-
-
-class BaseList(BaseModel):
-    """ Base pydantic class for lists of models """
-
-    def __iter__(self):
-        return iter(self.__root__)
-
-    def __contains__(self, value):
-        if isinstance(value, str):
-            return value in [i.name for i in self]
-        return value in self
-
-    def __getitem__(self, item):
-        if isinstance(item, str) and item in self:
-            vals = [i.name for i in self.__root__]
-            return self.__root__[vals.index(item)]
-        return self.__root__[item]
-
-    def __repr__(self):
-        val = '\n '.join(repr(i) for i in self)
-        return f"[{val}]"
-
-    def __str__(self):
-        return f"[{', '.join([i.name for i in self])}]"
-
-
+    
 class Releases(BaseList):
     """ Pydantic model representing a list of Releases """
     __root__: List[Release]